Refactor dataPoint
