+++ include/lib7842/api.hpp
@@ -21,8 +21,7 @@
 #include "lib7842/api/positioning/path/abstractPath.hpp"
 #include "lib7842/api/positioning/path/compoundPath.hpp"
 #include "lib7842/api/positioning/path/discretePath.hpp"
-#include "lib7842/api/positioning/point/dataPoint.hpp"
-#include "lib7842/api/positioning/point/dataState.hpp"
+#include "lib7842/api/positioning/point/data.hpp"
 #include "lib7842/api/positioning/point/mathPoint.hpp"
 #include "lib7842/api/positioning/point/state.hpp"
 #include "lib7842/api/positioning/point/vector.hpp"
