--- include/lib7842/api/positioning/point/dataPoint.hpp
@@ -1,56 +1,0 @@
-#pragma once
-#include "vector.hpp"
-
-#include <any>
-#include <map>
-#include <stdexcept>
-#include <string>
-
-namespace lib7842 {
-
-/**
- * A point that contains extra information of any type.
- */
-class DataPoint : public Vector {
-public:
-  using Vector::Vector;
-
-  /**
-   * Convert a Vector into a DataPoint
-   *
-   * @param ipoint The point
-   */
-  explicit DataPoint(const Vector& ipoint);
-
-  /**
-   * Set the point data.
-   *
-   * @param iid   The data name
-   * @param idata The data
-   */
-  void setData(const std::string& iid, const std::any& idata);
-
-  /**
-   * Get the point data.
-   *
-   * @param  iid The data name
-   * @tparam T   The data type
-   * @return The data
-   */
-  template <typename T> T getData(const std::string& iid) const {
-    const std::any& idata = getID(iid);
-    try {
-      return std::any_cast<T>(idata);
-    } catch (const std::bad_any_cast& e) {
-      throw std::runtime_error("DataPoint::getData:: \"" + iid + "\" contains wrong type \"" +
-                               idata.type().name() + "\"");
-    }
-  }
-
-protected:
-  std::map<std::string, std::any> data {};
-
-private:
-  const std::any& getID(const std::string& iid) const;
-};
-} // namespace lib7842
