--- include/lib7842/api/positioning/point/dataState.hpp
@@ -1,56 +1,0 @@
-#pragma once
-#include "state.hpp"
-
-#include <any>
-#include <map>
-#include <stdexcept>
-#include <string>
-
-namespace lib7842 {
-
-/**
- * A state that contains extra information of any type.
- */
-class DataState : public State {
-public:
-  using State::State;
-
-  /**
-   * Convert a State into a DataState
-   *
-   * @param istate The state
-   */
-  explicit DataState(const State& istate);
-
-  /**
-   * Set the data.
-   *
-   * @param iid   The data name
-   * @param idata The data
-   */
-  void setData(const std::string& iid, const std::any& idata);
-
-  /**
-   * Get the data.
-   *
-   * @param  iid The data name
-   * @tparam T   The data type
-   * @return The data
-   */
-  template <typename T> T getData(const std::string& iid) const {
-    const std::any& idata = getID(iid);
-    try {
-      return std::any_cast<T>(idata);
-    } catch (const std::bad_any_cast& e) {
-      throw std::runtime_error("DataState::getData:: \"" + iid + "\" contains wrong type \"" +
-                               idata.type().name() + "\"");
-    }
-  }
-
-protected:
-  std::map<std::string, std::any> data {};
-
-private:
-  const std::any& getID(const std::string& iid) const;
-};
-} // namespace lib7842
